% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_light_response_parameters.R
\name{calculate_light_response_parameters}
\alias{calculate_light_response_parameters}
\title{A function that calculates all light response curve parameters of a file.}
\usage{
calculate_light_response_parameters(
  pathname,
  manual_check = TRUE,
  save_plot = FALSE,
  save_path = "output_directory_licorfiles/light_response_plots/",
  name_parameters = c("date", "description", "light", "relative humidity", "CO2",
    "species", "measurement", "plant")
)
}
\arguments{
\item{pathname}{The licor file that will be analysed.}

\item{manual_check}{A boolean that regulates if the fits are checked.}

\item{name_parameters}{A character vector of the parameters in the filename.
Default is a list of date description, light,
relative humidity, CO2, species, measurement and plant.}
}
\value{
Returns a list with a list of the name parameters and a list of
the fitted light response parameters
}
\description{
A function that calculates all light response curve parameters of a file.
  The input is a pathname.
  It slits the path into directories and the filename.
  It splits the filename into name parameters.
  It makes a title for the plots.
  It fits the data using other functions.
  It returns a list of the list of name parameters and the list of fitted
  parameters.
}
\examples{
pathname <- "input_directory_licorfiles/light_response_data/
             20210226 LRC NA 75RH 400CO2 T.xlsx"
list_of_parameters <- calculate_light_response_parameters(pathname)

}
\author{
Sam Loontjens
}
