% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_light_response_curve.R
\name{fit_light_response_curve}
\alias{fit_light_response_curve}
\title{A function that fits a light response curve.}
\usage{
fit_light_response_curve(
  dataframe,
  title = "Light response curve",
  subtitle = "",
  manual_check = TRUE,
  save_plot = FALSE,
  save_path = "output_directory_licorfiles/light_response_plots/"
)
}
\arguments{
\item{dataframe}{The dataframe that will be analysed.}

\item{title}{A string for the title of the plot.}

\item{subtitle}{A string of the subtitle to use. Filename recommended.}

\item{manual_check}{A boolean that regulates if the fits are checked.}

\item{save_plot}{A boolean that regulates if the plots get saved.}

\item{save_path}{A string of the pathname where the plots get saved.}
}
\value{
Returns a list of the fitted light response parameters
}
\description{
A function that fits a light response curve.
  The input is a dataframe.
  It extracts the columns A, elapsed and Qin.
  (These have to be columns in the dataframe or it will result in an error)
  It sets the initial parameters Rd, alpha, Pmax, curvature and the
  light_compensation_point.
  If manual check is on it will plot the data.
  It will try to run the model.
  The formula is: -Rd + (alpha * PPF + Pmax - sqrt((alpha * PPF + Pmax)^2
  - 4 * curvature * PPF * Pmax * alpha))/(2 * curvature).
  If manual check is on it will ask the user if the fit is acceptable.
  If there is an error or warning it will return a list of NA values.
  If the fit is accepted it will return a list with the fitted parameters.
}
\examples{
fitted_parameters <- fit_light_response_curve(mydata)

}
\author{
Sam Loontjens
}
